% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_cases.R
\name{counts_to_cases}
\alias{counts_to_cases}
\title{Counts To Cases}
\usage{
counts_to_cases(df, countcol = "Freq")
}
\arguments{
\item{df}{Data frame of counts}

\item{countcol}{The name of the column containing the counts}
}
\value{
A data frame of cases
}
\description{
Convert from data frame of counts to data frame of cases. From:
https://cran.r-project.org/web/packages/DescTools/vignettes/TablesInR.pdf
}
